client/requirements.txt
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.2.1
Flask==1.1.2
Flask-HTTPAuth==4.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
opencv-python==4.4.0.46
pycparser==2.20
pynput==1.7.1
pyOpenSSL==19.1.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.25.0
six==1.15.0
updog==1.4
urllib3==1.26.2
websocket-client==0.57.0
Werkzeug==1.0.1
